+++ flattentool/cli.py
@@ -61,6 +61,10 @@ def create_parser():
         "--use-titles",
         action='store_true',
         help="Convert titles. Requires a schema to be specified.")
+    parser_create_template.add_argument(
+        "--disable-local-refs",
+        action='store_true',
+        help="Disable local refs when parsing JSON Schema.")
 
     parser_flatten = subparsers.add_parser(
         'flatten',
@@ -115,6 +119,11 @@ def create_parser():
     parser_flatten.add_argument(
         "--filter-value",
         help="Data Filter - only data with this will be processed. Use with --filter-field")
+    parser_flatten.add_argument(
+        "--disable-local-refs",
+        action='store_true',
+        help="Disable local refs when parsing JSON Schema.")
+
 
     parser_unflatten = subparsers.add_parser(
         'unflatten',
@@ -196,6 +205,10 @@ def create_parser():
         "--root-is-list",
         action='store_true',
         help="The root element is a list. --root-list-path and meta data will be ignored.")
+    parser_unflatten.add_argument(
+        "--disable-local-refs",
+        action='store_true',
+        help="Disable local refs when parsing JSON Schema.")
 
     return parser