+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Add --disable-local-refs to flatten, unflatten and create-template.
+
 ## [0.3.0] - 2018-10-12
 
 ### Added